Made significant progress
Implemented most classes into code. The driver file MUST be completed by tomorrow so we can finally test the application out and make a make-shift CLI GUI. We also need to verify that our UML design for the program works. On top of that, construct the test script and the Javadoc documentation.